// generated/src/app/api-models/pmb-sub-detail-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubDetailWork {
  seqGpbilId: number;
  seqSubDtlId: number;
  customerType: string;
  customerId: string;
  groupId: string;
  invoiceNo: number;
  seqSubsId: number;
  seqMembId: number;
  subscriberId: string;
  personNumber: string;
  employeeNumber: string;
  subscDept: string;
  subscLocation: string;
  billDetailFromDate: Date;
  billDetailThruDate: Date;
  recordType: string;
  planRiderCode: string;
  cobraFlag: string;
  billedEntityName: string;
  billedEntityRelationCode: string;
  billedEntityDateOfBirth: Date;
  billedEntityAge: number;
  billedEntityGender: string;
  familySize: number;
  premiumStep: string;
  monthlyRate: number;
  premiumAmt: number;
  salary: number;
  medicareStatus: string;
  seqParentId: number;
  rateType: string;
  spouseFlag: string;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
}